Animations working

+++ Headers.h
@@ -19,6 +19,10 @@
 // read docs: https://github.com/g-truc/glm/blob/master/manual.md#section1
 #include "glm/glm.hpp"
 #include "glm/ext.hpp"
+
+//#define GLM_ENABLE_EXPERIMENTAL
+//#include <glm/gtx/quaternion.hpp>
+
 #include <gl/GL.h>
 
 #include "stb_image.h"
@@ -55,7 +59,7 @@ struct vertexPolygon
 	glm::vec3 bitangent;
 
 	glm::vec4 weights;
-	glm::vec4 bones;
+	glm::ivec4 bones;
 };
 
 enum BOXTYPE 
@@ -64,7 +68,6 @@ enum BOXTYPE
 	EXPLOSIVE = 1,
 	ICEBLOCK = 2,
 	REGULAR = 3
-
 };
 enum ITEMTYPE 
 {
